+++ AddOns/Baganator/ItemViewCommon/Layouts.lua
@@ -934,6 +934,8 @@ function BaganatorLiveCategoryLayoutMixin:ShowGroup(cacheList, rowWidth, categor
       end
       if self.refreshContent then
         table.insert(toResetCache, {newButton, cacheData})
+      elseif newButton.BGR then
+        newButton.BGR.itemLocation = {bagID = cacheData.bagID, slotIndex = cacheData.slotID}
       end
     else
       if cacheData.isDummy then