+++ AddOns/Baganator/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Baganator
 
-## [618](https://github.com/Baganator/Baganator/tree/618) (2025-01-15)
-[Full Changelog](https://github.com/Baganator/Baganator/compare/617...618) 
+## [624](https://github.com/Baganator/Baganator/tree/624) (2025-01-18)
+[Full Changelog](https://github.com/Baganator/Baganator/compare/623...624) 
 
-- Show purchase price for new warband/guild bank tab in tooltip  
+- Fix priority not applying to items added to TSM (Auto) category